+++ util/src/main/java/tech/pegasys/teku/util/serialization/JsonExplicit.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.annotations;
+package tech.pegasys.teku.util.serialization;
 
 import com.fasterxml.jackson.annotation.JacksonAnnotationsInside;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
